Renamed some files, updated logo
Renamed some .scss and .js files to better describe their function. Made logo in navbar clickable and changed cursor to reflect that. When clicked it now takes you back to the home page. The whole site now loads with 'home.html' as the content from the start so as to avoid ugly flicker on page load. Changed title of browser tab to 'Benny's pizza'.